+++ snuba/admin/static/data.tsx
@@ -1,6 +1,7 @@
 import RuntimeConfig from "./runtime_config";
 import AuditLog from "./runtime_config/auditlog";
 import ClickhouseQueries from "./clickhouse_queries";
+import TracingQueries from "./tracing";
 
 function Placeholder(props: any) {
   return null;
@@ -13,6 +14,11 @@ const NAV_ITEMS = [
     id: "clickhouse",
     display: "ClickHouse🏚️",
     component: ClickhouseQueries,
+  },
+  {
+    id: "tracing",
+    display: "Tracing 🔎",
+    component: TracingQueries,
   },
   {
     id: "auditlog",
